// ErabliereIU/src/model/weatherforecast.ts
// Generated by https://quicktype.io

export class WeatherForecase {
    Headline?:       Headline;
    DailyForecasts?: DailyForecast[];
}

export interface DailyForecast {
    Date:        string;
    EpochDate:   number;
    Temperature: Temperature;
    Day:         Day;
    Night:       Day;
    Sources:     string[];
    MobileLink:  string;
    Link:        string;
}

export interface Day {
    Icon:                    number;
    IconPhrase:              string;
    HasPrecipitation:        boolean;
    PrecipitationType?:      string;
    PrecipitationIntensity?: string;
}

export interface Temperature {
    Minimum: Imum;
    Maximum: Imum;
}

export interface Imum {
    Value:    number;
    Unit:     Unit;
    UnitType: number;
}

export enum Unit {
    F = "F",
}

export interface Headline {
    EffectiveDate:      string;
    EffectiveEpochDate: number;
    Severity:           number;
    Text:               string;
    Category:           string;
    EndDate:            string;
    EndEpochDate:       number;
    MobileLink:         string;
    Link:               string;
}